+++ src/modules/files/c.mjs
@@ -1,0 +1,5 @@
+// console.log("Hello from c.js!");
+
+const consLog = console.log("Hello from c.js!");
+
+export default consLog;
